make checkIfPngsMatch shareable
