+++ x-pack/test/functional/apps/dashboard/reporting/screenshots.ts
@@ -6,14 +6,8 @@
  */
 
 import expect from '@kbn/expect';
-import fs from 'fs';
 import path from 'path';
-import { promisify } from 'util';
 import { FtrProviderContext } from '../../../ftr_provider_context';
-import { checkIfPngsMatch } from './lib/compare_pngs';
-
-const writeFileAsync = promisify(fs.writeFile);
-const mkdirAsync = promisify(fs.mkdir);
 
 const REPORTS_FOLDER = path.resolve(__dirname, 'reports');
 
@@ -27,6 +21,7 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
   const es = getService('es');
   const testSubjects = getService('testSubjects');
   const kibanaServer = getService('kibanaServer');
+  const reporting = getService('reporting');
   const ecommerceSOPath = 'x-pack/test/functional/fixtures/kbn_archiver/reporting/ecommerce.json';
 
   describe('Dashboard Reporting Screenshots', () => {
@@ -128,20 +123,6 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
     });
 
     describe('PNG Layout', () => {
-      const writeSessionReport = async (name: string, rawPdf: Buffer, reportExt: string) => {
-        const sessionDirectory = path.resolve(REPORTS_FOLDER, 'session');
-        await mkdirAsync(sessionDirectory, { recursive: true });
-        const sessionReportPath = path.resolve(sessionDirectory, `${name}.${reportExt}`);
-        await writeFileAsync(sessionReportPath, rawPdf);
-        return sessionReportPath;
-      };
-      const getBaselineReportPath = (fileName: string, reportExt: string) => {
-        const baselineFolder = path.resolve(REPORTS_FOLDER, 'baseline');
-        const fullPath = path.resolve(baselineFolder, `${fileName}.${reportExt}`);
-        log.debug(`getBaselineReportPath (${fullPath})`);
-        return fullPath;
-      };
-
       it('downloads a PNG file: small dashboard', async function () {
         this.timeout(300000);
 
@@ -155,10 +136,15 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
         const url = await PageObjects.reporting.getReportURL(60000);
         const reportData = await PageObjects.reporting.getRawPdfReportData(url);
         const reportFileName = 'small_dashboard_preserve_layout';
-        const sessionReportPath = await writeSessionReport(reportFileName, reportData, 'png');
-        const percentDiff = await checkIfPngsMatch(
+        const sessionReportPath = await PageObjects.reporting.writeSessionReport(
+          reportFileName,
+          'png',
+          reportData,
+          REPORTS_FOLDER
+        );
+        const percentDiff = await reporting.checkIfPngsMatch(
           sessionReportPath,
-          getBaselineReportPath(reportFileName, 'png'),
+          PageObjects.reporting.getBaselineReportPath(reportFileName, 'png', REPORTS_FOLDER),
           config.get('screenshots.directory'),
           log
         );
@@ -179,10 +165,15 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
         const url = await PageObjects.reporting.getReportURL(200000);
         const reportData = await PageObjects.reporting.getRawPdfReportData(url);
         const reportFileName = 'large_dashboard_preserve_layout';
-        const sessionReportPath = await writeSessionReport(reportFileName, reportData, 'png');
-        const percentDiff = await checkIfPngsMatch(
+        const sessionReportPath = await PageObjects.reporting.writeSessionReport(
+          reportFileName,
+          'png',
+          reportData,
+          REPORTS_FOLDER
+        );
+        const percentDiff = await reporting.checkIfPngsMatch(
           sessionReportPath,
-          getBaselineReportPath(reportFileName, 'png'),
+          PageObjects.reporting.getBaselineReportPath(reportFileName, 'png', REPORTS_FOLDER),
           config.get('screenshots.directory'),
           log
         );
