+++ x-pack/test/reporting_functional/services/scenarios.ts
@@ -6,8 +6,9 @@
  */
 
 import expect from '@kbn/expect';
-import { FtrProviderContext } from '../ftr_provider_context';
+import { checkIfPngsMatch } from '../../../../test/functional/services/lib/compare_pngs';
 import { createScenarios as createAPIScenarios } from '../../reporting_api_integration/services/scenarios';
+import { FtrProviderContext } from '../ftr_provider_context';
 
 export function createScenarios(
   context: Pick<FtrProviderContext, 'getPageObjects' | 'getService'>
@@ -161,5 +162,6 @@ export function createScenarios(
     tryReportsNotAvailable,
     loginDataAnalyst,
     loginReportingUser,
+    checkIfPngsMatch,
   };
 }
